+++ CHANGELOG.md
@@ -10,3 +10,8 @@ All notable changes to this project will be documented in this file.
 - Add state functionalities from `user`, `test-case` and `forecast-data`.
 - Sum the routing to the project.
 - Added `user` component to test dispatching the reducers.
+
+## [0.2.0] -- 2023-09-19
+
+- `Chart.js` package added in order to visualize graphs.
+- `chartjs-plugin-annotation` plugin appended in order to add box annotations on the fly.